Lots of cleanup.
open fails for reasons I don't understand.

We only have one warning left.

I'm beginning to think this might work ...

+++ sys/src/9/port/devsegment.c
@@ -477,8 +477,8 @@ static uintptr_t
 placeseg(uintptr_t len)
 {
 	static Lock lck;
-	static uint32_t va = HEAPTOP;
-	uint32_t v;
+	static uintptr_t va = HEAPTOP;
+	uintptr_t v;
 
 	len += BIGPGSZ;	/* so we fault upon overflows */
 	lock(&lck);
